dom/html/test/test_bug347174_write.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=347174
-->
<head>
  <title>Test for Bug 347174</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=347174">Mozilla Bug 347174</a>
<p id="display"></p>

<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 347174 **/
// simple test of readyState during loading, DOMContentLoaded, and complete
// this test passes in IE7
window.readyStateText = [];
window.loaded = false;
function attachCustomEventListener(element, eventName, command) {
    if (window.addEventListener && !window.opera)
        element.addEventListener(eventName, command, true);
    else if (window.attachEvent)
        element.attachEvent("on" + eventName, command);
}

function showMessage(msg) {
    window.readyStateText.push(msg);
    document.getElementById("display").innerHTML = readyStateText.join("<br>");
}

function frameLoad() {
  var doc = $('iframe').contentWindow.document;
  is(doc.readyState, "complete", "frame document.readyState should be 'complete' on load");
  showMessage("frame load: " + doc.readyState);
  if (window.loaded) SimpleTest.finish();
}

function load() {
    window.loaded = true;

    var imgsrc = "<img onload ='window.parent.imgLoad()' src='image.png'>\n";
    var doc = $('iframe').contentWindow.document;
    doc.writeln(imgsrc);
    doc.close();
    showMessage("frame after document.write: " + doc.readyState);
    isnot(doc.readyState, "complete", "frame document.readyState should not be 'complete' after document.write");
}

function imgLoad() {
    var doc = $('iframe').contentWindow.document;
    showMessage("frame after imgLoad: " + doc.readyState);
    is(doc.readyState, "interactive", "frame document.readyState should still be 'interactive' after img loads");
}

window.onload=load;

SimpleTest.waitForExplicitFinish();

</script>
</pre>
<iframe src="404doesnotexist" id="iframe" onload="frameLoad();"></iframe>
</body>
</html>
